feat(discount-codes): update code according to review
affects: @commercetools/discount-code-generator
