+++ packages/discount-code-generator/src/main.js
@@ -5,33 +5,35 @@ import type {
   CodeData,
   CodeDataArray,
   CodeOptions,
-} from 'types/sdk'
+} from 'types/discountCodes'
 
-// The discountCodeGenerator function takes 2 arguments {options} and {data}
-// The {options} will specify the semantics and quantity of discount codes
-// Will throw if no quantity `number` is specified
-// Defaults to 11-digit codes if no length `number` is specified
-// Can also specify an optional prefix `string`
-// options = {
-//   quantity: 100,
-//   length: 15,
-//   prefix: CT,
-// }
-//
-// The {data} should have the attributes of the discount codes
-// data = {
-//   name: {
-//    en: 'foo'
-//   },
-//   description: {
-//    en: 'bar'
-//   },
-//   cartDiscounts: [],
-//   cartPredicate: 'some predicate',
-//   isActive: true,
-//   maxApplicationsPerCustomer: 10,
-//   maxApplicationsPerCustomer: 2,
-//  }
+/**
+ * The discountCodeGenerator function takes 2 arguments {options} and {data}
+ * The {options} will specify the semantics and quantity of discount codes
+ * Will throw if no quantity `number` is specified
+ * Defaults to 11-digit codes if no length `number` is specified
+ * Can also specify an optional prefix `string`
+ * options = {
+ *   quantity: 100,
+ *   length: 15,
+ *   prefix: CT,
+ * }
+
+ * The {data} should have the attributes of the discount codes
+ * data = {
+ *   name: {
+ *    en: 'foo'
+ *   },
+ *   description: {
+ *    en: 'bar'
+ *   },
+ *   cartDiscounts: [],
+ *   cartPredicate: 'some predicate',
+ *   isActive: true,
+ *   maxApplicationsPerCustomer: 10,
+ *   maxApplicationsPerCustomer: 2,
+ *  }
+**/
 
 export default function discountCodeGenerator (
   options: CodeOptions,
@@ -43,7 +45,7 @@ export default function discountCodeGenerator (
     throw new Error('The generator requires discount data')
   const { length, prefix, quantity } = { length: 11, prefix: '', ...options }
   const codes = []
-  const chars = '0-9A-Z'
+  const chars = '0-9a-zA-Z'
   const generator = new TokenGenerator({ chars, length })
   for (let i = 0; i < quantity; i += 1) {
     const codeObject = Object.assign({}, data)
@@ -57,6 +59,5 @@ function _prepareCode (code: string, length: number, prefix: string): string {
   if (!prefix.length)
     return code
 
-  const codePrefix = prefix.toUpperCase()
-  return `${codePrefix}${code.slice(-(length - codePrefix.length))}`
+  return `${prefix}${code.slice(-(length - prefix.length))}`
 }
